add ability to set registry, repository and tag values and support private registries (#11)
Particular enterprise customers may require that images be pulled from
private registries that require authentication.

This PR allows that.

This PR also moves away from prefixing tags with `prod-` which seemed to
be an intentional choice but (my opinion) would be worth review. I
_guess_ anyone wanting to utilize an image that they have tagged
themselves _could_ prefix their own images with `-prod` but it would be
unlikely they'd want to.