Nango Helm Chart V2 (#25)
Major rework of the Nango helm chart - now on v2. 

This is heavily inspired by the [bitnami
template](https://github.com/bitnami/charts/tree/main/template).

It supports a range of features for each component and can be easily
configured from the values file. Sensible defaults are provided.

The chart does not provide _any_ of the application environment
variables - all of these are provided by the consumer application.

Next, will work on a better README and documentation.